Update to LLVM r172983.
git-svn-id: http://ellcc.org/svn/ellcc/trunk@3185 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ llvm/tools/clang/lib/CodeGen/CGRecordLayoutBuilder.cpp
@@ -21,11 +21,11 @@
 #include "clang/AST/Expr.h"
 #include "clang/AST/RecordLayout.h"
 #include "clang/Frontend/CodeGenOptions.h"
-#include "llvm/DataLayout.h"
-#include "llvm/DerivedTypes.h"
+#include "llvm/IR/DataLayout.h"
+#include "llvm/IR/DerivedTypes.h"
+#include "llvm/IR/Type.h"
 #include "llvm/Support/Debug.h"
 #include "llvm/Support/raw_ostream.h"
-#include "llvm/Type.h"
 using namespace clang;
 using namespace CodeGen;
 
@@ -1075,7 +1075,8 @@ CGRecordLayout *CodeGenTypes::ComputeRecordLayout(const RecordDecl *D,
       // the size + offset should match the storage size in that case as it
       // "starts" at the back.
       if (getDataLayout().isBigEndian())
-        assert((Info.Offset + Info.Size) == Info.StorageSize &&
+        assert(static_cast<unsigned>(Info.Offset + Info.Size) ==
+               Info.StorageSize &&
                "Big endian union bitfield does not end at the back");
       else
         assert(Info.Offset == 0 &&
